requirements.txt
diffusers==0.34.0
grad_cam==1.5.5
matplotlib==3.10.3
numpy==2.3.2
ollama==0.5.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.1
Pillow==11.3.0
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.16.0
seaborn==0.13.2
skimage==0.0
tensorflow==2.16.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
ultralytics==8.3.163
